requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
pandas==2.0.3
reportlab==4.0.4
